CI: Add workflow to update Pipfile.lock every month
Adds a workflow that generates a up to date lockfile based on the current Pipfile. The workflow runs every 1st day of each month, and if either this configuration or the Pipfile itself is modified. A run will open a PR with the updated pipfile.lock, and if a PR is already open it will update that PR.